+++ pkg/api/handlers.go
@@ -276,7 +276,7 @@ func NewQuerierHandler(
 		// This is used for the stats API which we should not support. Or find other ways to.
 		prometheus.GathererFunc(func() ([]*dto.MetricFamily, error) { return nil, nil }),
 		reg,
-		nil,
+		querier.StatsRenderer,
 		remoteWriteEnabled,
 		oltpEnabled,
 	)