+++ pkg/frontend/transport/handler_test.go
@@ -65,13 +65,14 @@ func TestHandler_ServeHTTP(t *testing.T) {
 	const testRouteName = "the_test_route"
 
 	for _, tt := range []struct {
-		name                    string
-		cfg                     HandlerConfig
-		request                 func() *http.Request
-		expectedParams          url.Values
-		expectedMetrics         int
-		expectedActivity        string
-		expectedReadConsistency string
+		name                     string
+		cfg                      HandlerConfig
+		request                  func() *http.Request
+		expectedParams           url.Values
+		expectedMetrics          int
+		expectedActivity         string
+		expectedReadConsistency  string
+		expectedExtraStatsHeader string
 	}{
 		{
 			name: "handler with stats enabled, POST request with params",
@@ -90,9 +91,10 @@ func TestHandler_ServeHTTP(t *testing.T) {
 				"query": []string{"some_metric"},
 				"time":  []string{"42"},
 			},
-			expectedMetrics:         5,
-			expectedActivity:        "user:12345 UA:test-user-agent req:POST /api/v1/query query=some_metric&time=42",
-			expectedReadConsistency: "",
+			expectedMetrics:          5,
+			expectedActivity:         "user:12345 UA:test-user-agent req:POST /api/v1/query query=some_metric&time=42",
+			expectedReadConsistency:  "",
+			expectedExtraStatsHeader: "total_samples=0",
 		},
 		{
 			name: "handler with stats enabled, GET request with params",
@@ -106,9 +108,10 @@ func TestHandler_ServeHTTP(t *testing.T) {
 				"query": []string{"some_metric"},
 				"time":  []string{"42"},
 			},
-			expectedMetrics:         5,
-			expectedActivity:        "user:12345 UA:test-user-agent req:GET /api/v1/query query=some_metric&time=42",
-			expectedReadConsistency: "",
+			expectedMetrics:          5,
+			expectedActivity:         "user:12345 UA:test-user-agent req:GET /api/v1/query query=some_metric&time=42",
+			expectedReadConsistency:  "",
+			expectedExtraStatsHeader: "total_samples=0",
 		},
 		{
 			name: "handler with stats enabled, GET request with params and read consistency specified",
@@ -122,9 +125,10 @@ func TestHandler_ServeHTTP(t *testing.T) {
 				"query": []string{"some_metric"},
 				"time":  []string{"42"},
 			},
-			expectedMetrics:         5,
-			expectedActivity:        "user:12345 UA:test-user-agent req:GET /api/v1/query query=some_metric&time=42",
-			expectedReadConsistency: api.ReadConsistencyStrong,
+			expectedMetrics:          5,
+			expectedActivity:         "user:12345 UA:test-user-agent req:GET /api/v1/query query=some_metric&time=42",
+			expectedReadConsistency:  api.ReadConsistencyStrong,
+			expectedExtraStatsHeader: "total_samples=0",
 		},
 		{
 			name: "handler with stats enabled, GET request without params",
@@ -134,10 +138,11 @@ func TestHandler_ServeHTTP(t *testing.T) {
 				r.Header.Add("User-Agent", "test-user-agent")
 				return r
 			},
-			expectedParams:          url.Values{},
-			expectedMetrics:         5,
-			expectedActivity:        "user:12345 UA:test-user-agent req:GET /api/v1/query (no params)",
-			expectedReadConsistency: "",
+			expectedParams:           url.Values{},
+			expectedMetrics:          5,
+			expectedActivity:         "user:12345 UA:test-user-agent req:GET /api/v1/query (no params)",
+			expectedReadConsistency:  "",
+			expectedExtraStatsHeader: "total_samples=0",
 		},
 		{
 			name: "handler with stats disabled, GET request with params",
@@ -191,6 +196,12 @@ func TestHandler_ServeHTTP(t *testing.T) {
 			handler.ServeHTTP(resp, req)
 			responseData, _ := io.ReadAll(resp.Body)
 			require.Equal(t, resp.Code, http.StatusOK)
+
+			if tt.expectedExtraStatsHeader == "" {
+				assert.NotContains(t, resp.Header(), "X-Mimir-Query-Stats")
+			} else {
+				assert.Equal(t, tt.expectedExtraStatsHeader, resp.Header().Get("X-Mimir-Query-Stats"))
+			}
 
 			count, err := promtest.GatherAndCount(
 				reg,